refactor: move unfold widgets to their own package
