+++ src/widgets/__init__.py
@@ -1,11 +1,11 @@
 from .config import UNFOLD_BUTTON_WIDTH
 from .table_model import QtTable
-from .unfold_widget import UnfoldWidget
-from .algorithm_setup_widget import AlgorithmSetupWidget
-from .algorithm_run_widget import AlgorithmRunWidget
-from .import_widget import ImportWidget
-from .preprocessing_widget import PreprocessingWidget
-from .results_widget import ResultsWidget
+from .unfold_widgets.unfold_widget import UnfoldWidget
+from .unfold_widgets.algorithm_setup_widget import AlgorithmSetupWidget
+from .unfold_widgets.algorithm_run_widget import AlgorithmRunWidget
+from .unfold_widgets.import_widget import ImportWidget
+from .unfold_widgets.preprocessing_widget import PreprocessingWidget
+from .unfold_widgets.results_widget import ResultsWidget
 from .main_widget import MainWidget
 from .main_window import MainWindow
 from .plots import HistogramPlot, PiePlot, FallbackPlot
