+++ src/widgets/unfold_widgets/algorithm_setup_widget.py
